Creation of the maintenance and update microservice for the utils project

+++ Train/src/main/java/br/com/railvista/RailVista/Service/TrainService.java
@@ -1,9 +1,12 @@
 package br.com.railvista.RailVista.Service;
 
-import br.com.railsUtils.TrainDTO;
-import br.com.railsUtils.TrainNotFoundException;
+import br.com.railsUtils.DTO.MaintenanceDTO;
+import br.com.railsUtils.DTO.TrainDTO;
+import br.com.railsUtils.Exceptions.TrainInMaintenanceException;
+import br.com.railsUtils.Exceptions.TrainNotFoundException;
 import br.com.railvista.RailVista.Entities.Train;
 import br.com.railvista.RailVista.Repository.TrainRepository;
+import br.com.railvista.RailVista.Usecase.MaintenanceService;
 import br.com.railvista.RailVista.Utilis.DTOConverter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,12 +14,13 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 public class TrainService {
     @Autowired
     private TrainRepository trainRepository;
+    @Autowired
+    private MaintenanceService maintenanceService;
 
 
     public TrainDTO save(TrainDTO trainDTO) {
@@ -34,11 +38,25 @@ public class TrainService {
 
     public TrainDTO deleteById(UUID id) {
         Optional<Train> train = trainRepository.findById(id);
-        trainRepository.deleteById(id);
+
+        if (maintenanceService.getMaintenanceByIdTrain(id) == null) {
+            if (train.isEmpty()) {
+                throw new TrainNotFoundException();
+            }
+            if (train.isPresent()) {
+                trainRepository.deleteById(id);
+            }
+        } else {
+            throw new TrainInMaintenanceException();
+        }
         return DTOConverter.convert(train.get());
     }
 
-    public void deleteAll() {
-        trainRepository.deleteAll();
+    public TrainDTO findById(UUID id) {
+        Optional<Train> train = trainRepository.findById(id);
+        if (train.isEmpty()){
+            throw new TrainNotFoundException();
+        }
+        return DTOConverter.convert(train.get());
     }
 }
